+++ projects/Layer - Util/teal/src/messages/raft.rs
@@ -2,7 +2,7 @@ use crate::{net::message::MessageIdentifiable, protos::gen::raft::Heartbeat};
 
 
 impl MessageIdentifiable for Heartbeat {
-    fn id(&self) -> u8 {
+    fn id(&self) -> u16 {
         11
     }
 }
